Change how packet handling is done.
Separate handler classes allow for easier data management.

On the client side it is no longer necessary to keep track of job and status data structures and associate incomming packets with the correct structures.  The handler classes will manage the structures and any related packets.

It is also easier to know when the client can disconnect from the server as no more handlers will be attached.  Disconnecting from the server as soon as possible is necessary to let the main loop and thus script end.